Fixed issues with theme
This was a mess. Setting default theme to dark enable system to false
in next-themes may override the device's theme, but the browser still
forces its own theme, and basically inverts everything so that it's a
total mess. The way I found to fix this is to ALWAYS specify the colors
you want to use for a component so that it has no choice but to follow
those. If nothing is specified, it will do whatever it wants